pyproject.toml
[project]
name = "print-without-black"
version = "0.1.0"
description = ""
authors = [
    {name = "ThanAid",email = "thanos.aidinis@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pdf2image (>=1.17.0,<2.0.0)",
    "pillow (>=12.1.0,<13.0.0)",
    "streamlit (>=1.53.0,<2.0.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "ruff (>=0.14.11,<0.15.0)"
]

[tool.ruff]
# Same as Black.
line-length = 88
target-version = "py310"

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`)  codes by default.
# Unlike Flake8, Ruff doesn't enable pycodestyle warnings (`W`) or
# McCabe complexity (`C901`) by default.
select = ["E4", "E7", "E9", "F"]
ignore = []

# Allow fix for all enabled rules (when `--fix` is provided).
fixable = ["ALL"]
unfixable = []